Staging

+++ includes/class-settings.php
@@ -39,7 +39,7 @@ class Settings {
 	public static function get( $var = null ) {
 		if ( $var ) {
 			$field = self::get_field( $var );
-			if ( $field['value'] ) {
+			if ( isset( $field['value'] ) ) {
 				return $field['value'];
 			} elseif ( isset( self::$settings[ $var ] ) ) {
 				return apply_filters( 'murmurations_aggregator_get_setting', self::$settings[ $var ], $var );
@@ -72,10 +72,10 @@ class Settings {
 		$schema_fields = self::get_fields();
 
 		foreach ( $schema_fields as $field => $attribs ) {
-			if ( $attribs['value'] ) {
+			if ( isset( $attribs['value'] ) ) {
 				$settings[ $field ] = $attribs['value'];
 			} elseif ( $attribs['default'] ) {
-				if ( $settings[ $field ] === null || $settings[ $field ] === '' || ( $settings[ $field ] === false && $attribs['type'] !== 'boolean' ) ) {
+				if ( null === $settings[ $field ] || '' === $settings[ $field ] || ( false === $settings[ $field ] && 'boolean' !== $attribs['type'] ) ) {
 					$settings[ $field ] = $attribs['default'];
 				}
 			}
@@ -94,7 +94,12 @@ class Settings {
 	 */
 	public static function save() {
 		$settings = apply_filters( 'murmurations_aggregator_save_settings', self::$settings );
-		return update_option( 'murmurations_aggregator_settings', $settings );
+		$existing = get_option( 'murmurations_aggregator_settings' );
+		if ( $existing !== $settings ) {
+			return update_option( 'murmurations_aggregator_settings', $settings );
+		} else {
+			return true;
+		}
 	}
 
 	/**
@@ -103,6 +108,7 @@ class Settings {
 	 * @return string Schema JSON
 	 */
 	private static function get_schema_json() {
+		// phpcs:ignore WordPress.WP.AlternativeFunctions.file_get_contents_file_get_contents
 		return file_get_contents( MURMAG_ROOT_PATH . 'admin_fields_jschema.json' );
 	}
